+++ uCOS-II_Wrapper/uCOS-II/os_cfg.h
@@ -124,7 +124,7 @@
 #else
 #define OS_MUTEX_EN               0u   /* Enable (1) or Disable (0) code generation for MUTEX          */
 #endif
-#define OS_MUTEX_ACCEPT_EN        0u   /*     Include code for OSMutexAccept()                         */
+#define OS_MUTEX_ACCEPT_EN        1u   /*     Include code for OSMutexAccept()                         */
 #define OS_MUTEX_DEL_EN           1u   /*     Include code for OSMutexDel()                            */
 #define OS_MUTEX_QUERY_EN         1u   /*     Include code for OSMutexQuery()                          */